#Fix: change int to long for longer numbers

+++ f_integer.c
@@ -6,8 +6,8 @@
  */
 int f_integer(va_list char_list)
 {
-	int num = va_arg(char_list, int);
-	int len = 0;
+	long num = va_arg(char_list, long);
+	long len = 0;
 
 	if (num < 0)
 	{
@@ -23,9 +23,9 @@ int f_integer(va_list char_list)
  * @num: num to be print
  * Return: the lenght of the string in abs. value
  */
-int print_integer(int num)
+int print_integer(long num)
 {
-	int len = 0;
+	long len = 0;
 
 	if (num / 10)
 	{
